+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+- fix: guard `JSON.parse` in dest_domains and force_routing
+- test: refactored against test-fixtures 1.7.0
+
 ### [1.0.2] - 2026-05-15
 
 - feat: allow comments in relay_acl_allow